refactor: loggers (#323)
* ref: move elasticaLogger to elasticSearch namespace

* ref: remove unused ElasticsearchLogger.php

* ref: remove flashMessageLogger (move core)

* phpstan
